TST/DOC: add keyword and keyword-problem test and documentation
