+++ ramp-frontend/frontend/views/__init__.py
@@ -1,0 +1,2 @@
+"""The :mod:`frontend.views` defines all operations a user can need to deal
+with the RAMP backend."""
